feat: shop ui almost finished

+++ databaseProject/src/main/java/InterfaceImplementation/PriceInterface.java
@@ -20,13 +20,13 @@ import static InterfaceImplementation.CommodityInterface.getCommodityPriceHistor
  * @date 2023-12-02
  **/
 public class PriceInterface {
-    public static Integer insertPrice(Connection conn, Integer cId, Integer price) {
+    public static Integer insertPrice(Connection conn, Integer cId, Double price) {
         Integer result = -1;
         try{
             String sql = "insert into price(c_id, price, time) values(?, ?, current_timestamp)";
             PreparedStatement ps = conn.prepareStatement(sql);
             ps.setInt(1, cId);
-            ps.setInt(2, price);
+            ps.setDouble(2, price);
             result =  ps.executeUpdate();
         } catch (Exception e) {
             e.printStackTrace();
